:art: Move unit test fixtures under `tests/src/unit/conftest.py`
